requirements.txt
aiohttp==3.8.5
asyncio==3.4.3
clickhouse-driver==0.2.6
pydantic==2.3.0
python-dotenv==1.0.0
tenacity==8.2.3
pyyaml==6.0.1
pytest==7.4.2
pytest-asyncio==0.21.1
clickhouse-connect==0.8.15
pyarrow==19.0.1